Fix more bugs around check_any
This never seems to have really worked? At any rate, this needs to adjust
the Tat by one fewer element than requested; the weight needs to be added
before saturatingly subtracting Tau (otherwise the boundary is in the wrong
position), and in the end, we have to add a `t`. Welpppp.

Also, add a test to ensure that the regular check works the same as
check_all(1).