Fix embarrassing bug in NotUntil calculation
The earliest possible time a cell might be let through is:
theoretical next arrival time, less the burst size of the bucket.

I got this entirely wrong, only returning the tat.

In order to not get it wrong in the future, add a proptest that
exercises buckets of varying fill grades.