Make the 'nanos' module public
As it stands, it's impossible for a library consumer to supply an
alternative implementation of `Clock`, as the trait definition for
`Reference` requires that measurements be compared in terms of `Nanos`,
which is defined in a private module.

Since the rest of the library is clearly designed to allow the consumer
to bring their own clock, this looks like it was just an oversight.

Also adds a short doctest example to `clock.rs` to ensure the contract
is upheld in future.