Change the API for check_n / check_n_keyed / until_n / until_n_keyed
Previously, it returned an enum whose variants had different jobs: one
to indicate that a usage error had occurred and one to indicate the
rate limiting result. That's not ideal as it complicated every use of
the rate limiter.

Instead, return a nested result: The outermost can be used with the
try/? operator, to ensure usage errors get propagated through, and
then the one can be used like you would a non-`n` usage of the rate
limiter.