Differentiate between GCRA burst capacity and tolerance
In GCRA terms, the total capacity of the rate limiter is `Τ + τ`, where
`Τ` is the nominal cell arrival interval and `τ` is (extra) tolerance
allowed for bursts.

The current implementation has a subtle issue, caused by conflating
capacity and tolerance, which results in one extra cell being allowed
after arrival intervals equal or greater than "capacity + 1", as
reported in #107 and #249.

This problem was partially hidden by a compensating addition of Τ to
initial TAT in the case of the first cell (i.e. when the rate limiter
has no state recorded for it).

To fix the issue, correct the value `tau` used for making the decision,
making it the tolerance, and remove the adjustment for the first cell
that's no longer necessary.

Fixes: #107
Fixes: #249